More asserts
This version with in-band headers gets an assertion in MPS. Nailboard
is called with a client address that is inside [client_base
client_limit), but the Nailboard seems to have base addresses. With a
suitale address that is == base_limit, this leads to the assertion.
This let's me suspect that the in-band header part should maybe not be
used.