Correct for some incorrect pseudoheader sizes in igc.c
Arguably, these are bugs that should be fixed in lisp.h (along with the
comment describing the pvec header for subrs) and thread.c, by
initializing pseudovector headers so they actually describe the
pseudovectors in static memory, but the traditional GC doesn't care.

* src/alloc.c (make_pure_bignum): Call 'igc_init_header' after setting
the pvec header, not before.
* src/igc.c (gc_init_header): Catch PVEC_SUBR and PVEC_THREAD, which
sometimes have incorrect headers
* src/lisp.h (DEFUN) [HAVE_MPS]: Set pseudovector flag for subrs.