Change terminal cursor display in child frames
Frame parameter tty-cursor is gone. Frame parameter cursor-type
determines if a cursor is displayed in a child frame if it is not the
selected frame.  Set that for Vertico-posframe to put the terminal's
cursor in the posframe.

* src/dispnew.c (is_cursor_obscured): New function.
(terminal_cursor_magic): New function.
(combine_updates_for_frame): Call  terminal_cursor_magic.
(syms_of_display): Remove unused DEFSYMs.
* src/term.c (tty_show_cursor): Make public.
* src/termhooks.h: Declare it.