Don't allow buffer access from dying threads
This seems to happen when MacOS tries to invoke an 'autocompletion'
which causes Emacs to read from the current buffer. The problem with
this is that MacOS can try to do this at random times, including on a
thread that is currently dying.

Perhaps it's more appropriate to fix these handlers to defer to the main
thread somehow, but this fix appears to work.