Prevent warnings about incompatible function types in mac.c
these warnings are actually correct, the mac functions don not
have the same signature as the framework functions, although
they are compatible. For example, they DescType arguments where
unsigned int is expected, and long for the handlerReCon which is
actualy a SRefCon, which is a void *.

It's too much work for me to change all callers etc., so...

* src/mac.c (init_coercion_handler): Use a pragma to prevent the
warnings.