Remove excessive CF_NOESCAPE and NS_NOESCAPE from mac
These annotations direct the compiler to optimize block allocation
assuming the function to which it is passed calls it locally, and does
not retain the block beyond the function body.

Analyzed all occurrences of CF_NOESCAPE and NS_NOESCAPE and removed
locations where the annotation seemed to be excessive, for example when
the annotated block has been enqueued in a queue to execute in another
loop.

Such a block may have a an optimized placement as
`NSConcreteGlobalBlock' on the stack on a callee side. This happen on
non-Apple clang Apple at least from version 7 (see [1]).

This tries to preserve as many of the NOESCAPES which serve a purpose;
see [2], [3], and [4].

* src/macappkit.m (mac_within_gui, mac_within_lisp,
  mac_within_gui_and_here, mac_within_gui_allowing_inner_lisp): Remove
  excessive CF_NOESCAPE annotations, for blocks that are passed and
  executed in other threads or saved on queues.
* src/macterm.h (mac_within_gui) : Remove excessive CF_NOESCAPE, for
  blocks that are passed and executed in other threads.
* src/macappkit.h (enumerateChildWindowsUsingBlock): Harmonize
  NS_NOESCAPE placement.
* src/macappkit.m (enumerateChildWindowsUsingBlock): Harmonize
  CF_NOESCAPE placement.

[1] https://github.com/NixOS/nixpkgs/issues/127902#issuecomment-1698365843
[2] https://github.com/jdtsmith/emacs-mac/issues/43
[3] https://github.com/jdtsmith/emacs-mac/pull/48
[4] https://github.com/jdtsmith/emacs-mac/issues/69